ai-workflows/requirements.txt
prefect>=2.14.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
requests>=2.0.0
matplotlib>=3.0.0
torch>=2.0.0 